Add retry logic for TransferSettlementStatus job (#37658)
* Add retry logic for TransferSettlementStatus job

* Update tracecode

* Add testcases

* Minor change

* Skip settlement status update for transferID if it is already set as settled

* Update test

---------

+++ app/Jobs/Transfers/TransferSettlementStatus.php
@@ -8,9 +8,16 @@ use RZP\Jobs\Job;
 use RZP\Trace\TraceCode;
 use Razorpay\Trace\Logger;
 use RZP\Models\Transfer\Core;
+use RZP\Exception\SettlementStatusUpdateException;
 
 class TransferSettlementStatus extends Job
 {
+    const DISPATCH_DELAY_SECONDS = 300;
+
+    const RETRY_INTERVAL = 300;
+
+    const MAX_RETRY_ATTEMPT = 3;
+
     /**
      * This job will be terminated after <$timeout> seconds.
      * @var int
@@ -42,21 +49,52 @@ class TransferSettlementStatus extends Job
         try
         {
             (new Core())->updateSettlementStatusInTransfers($this->settlementId);
+
+            $this->delete();
         }
         catch (\Exception $ex)
         {
-            $this->trace->traceException(
-                $ex,
-                Logger::ERROR,
-                TraceCode::SETTLEMENT_STATUS_UPDATE_IN_TRANSFER_FAILED,
+            if ($ex instanceof SettlementStatusUpdateException)
+            {
+                $this->checkRetry();
+            }
+            else
+            {
+                $this->trace->traceException(
+                    $ex,
+                    Logger::ERROR,
+                    TraceCode::SETTLEMENT_STATUS_UPDATE_IN_TRANSFER_FAILED,
+                    [
+                        'settlement_id' => $this->settlementId,
+                    ]
+                );
+            }
+        }
+    }
+
+    protected function checkRetry()
+    {
+        if ($this->attempts() > self::MAX_RETRY_ATTEMPT)
+        {
+            $this->trace->error(
+                TraceCode::SETTLEMENT_STATUS_UPDATE_RETRY_EXHAUSTED,
                 [
                     'settlement_id' => $this->settlementId,
                 ]
             );
-        }
-        finally
-        {
+
             $this->delete();
+        }
+        else
+        {
+            $this->trace->info(
+                TraceCode::SETTLEMENT_STATUS_UPDATE_RETRY_DISPATCH,
+                [
+                    'settlement_id' => $this->settlementId,
+                ]
+            );
+
+            $this->release(self::RETRY_INTERVAL);
         }
     }
 }
